+++ cmd/src/batch_common.go
@@ -287,7 +287,10 @@ func executeBatchSpec(ctx context.Context, ui ui.ExecUI, opts executeBatchSpecOp
 
 	if len(batchSpec.Steps) > 0 {
 		ui.PreparingContainerImages()
-		images, err := svc.EnsureDockerImages(ctx, batchSpec.Steps, ui.PreparingContainerImagesProgress)
+		images, err := svc.EnsureDockerImages(
+			ctx, batchSpec.Steps, opts.flags.parallelism,
+			ui.PreparingContainerImagesProgress,
+		)
 		if err != nil {
 			return err
 		}