+++ internal/batches/docker/cache.go
@@ -2,15 +2,19 @@ package docker
 
 import "sync"
 
-// ImageCache is a cache of metadata about Docker images, indexed by name.
-type ImageCache struct {
+type ImageCache interface {
+	Get(name string) Image
+}
+
+// imageCache is a cache of metadata about Docker images, indexed by name.
+type imageCache struct {
 	images   map[string]Image
 	imagesMu sync.Mutex
 }
 
 // NewImageCache creates a new image cache.
-func NewImageCache() *ImageCache {
-	return &ImageCache{
+func NewImageCache() ImageCache {
+	return &imageCache{
 		images: make(map[string]Image),
 	}
 }
@@ -18,7 +22,7 @@ func NewImageCache() *ImageCache {
 // Get returns the image cache entry for the given Docker image. The name may be
 // anything the Docker command line will accept as an image name: this will
 // generally be IMAGE or IMAGE:TAG.
-func (ic *ImageCache) Get(name string) Image {
+func (ic *imageCache) Get(name string) Image {
 	ic.imagesMu.Lock()
 	defer ic.imagesMu.Unlock()