+++ CHANGELOG.md
@@ -13,6 +13,16 @@ All notable changes to `src-cli` are documented in this file.
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Removed
+
+## 3.36.4
+
+### Added
+
 - A new `src batch remote` command has been added to run a batch spec on a Sourcegraph instance that has Server Side Batch Changes enabled. [#671](https://github.com/sourcegraph/src-cli/pull/671)
 
 ### Changed
@@ -22,8 +32,6 @@ All notable changes to `src-cli` are documented in this file.
 ### Fixed
 
 - A bug in 3.36.3 caused `src batch apply` to no longer accept a file provided with `-f`. This has been fixed. [#695](https://github.com/sourcegraph/src-cli/pull/695)
-
-### Removed
 
 ## 3.36.3